back/pyproject.toml
[tool.poetry]
name = "vineguard-backend"
version = "0.1.0"
description = "Backend for VineGuard vineyard management system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.23.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
tortoise-orm = "^0.19.0"
asyncpg = "^0.25.0"
aerich = "^0.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
websockets = "^11.0.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
email-validator = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.0"

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 